"fix: optimize reuse function for leaflet modes"
