+++ src/measure-modes/leaflet/TwoPointsDistanceLeaflet.js
@@ -68,6 +68,21 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
         },
     };
 
+    #polylineListeners = {
+        mousedown: (polyline, event) => {
+            if (event.domEvent.button === 1) {
+                // Prevent map drag, default behavior
+                event.domEvent?.stopPropagation();
+                event.domEvent?.preventDefault();
+                // When the measure is completed or not started yet, make it interactive
+                // Switch mode case: isMeasurementComplete flags is not used (false) when at the beginning before the measure starts
+                if (this.coordsCache.length === 0) {
+                    // Handle polyline click logic here, if needed
+                    this._removeLineSet(polyline);
+                }
+            }
+        }
+    }
 
     /**
      * 
@@ -150,23 +165,19 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
         dataPool.updateOrAddMeasure({ ...this.measure });
 
         if (this.coordsCache.length === 2) {
-            // update status pending annotations
-            const pointsArray = this.pointCollection.getLayers();
-            pointsArray.forEach(point => {
-                if (point.id?.includes(`annotate_${this.mode}`) || point?.feature?.properties?.status) {
-                    point.feature.properties.status = "completed";
-                }
-            });
+            // -- Update annotations status --
+            // Update points status and interactive
+            this._updatePendingItemsToCompleted(this.pointCollection.getLayers(), `annotate_${this.mode}`);
 
-            // -- APPROACH 2: Update/ Reuse existing polyline and label --
-            // -- Handle polyline --
+            // -- APPROACH 1: Remove and recreate polyline --
             this._createOrUpdateLine(this.coordsCache, this.#interactiveAnnotations.polylines, {
                 status: "completed",
                 color: this.stateManager.getColorState("line"),
-                interactive: true
+                interactive: true,
+                listeners: this.#polylineListeners
             });
 
-            // -- Handle label --
+            // -- APPROACH 2: Update/ Reuse existing label --
             const { distance } = this._createOrUpdateLabel(this.coordsCache, this.#interactiveAnnotations.labels, {
                 status: "completed",
                 interactive: true,
@@ -215,7 +226,7 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
                 this._createOrUpdateLine(positions, this.#interactiveAnnotations.polylines, {
                     status: "moving",
                     color: this.stateManager.getColorState("move"),
-                    interactive: false
+                    interactive: false,
                 });
 
                 // Moving label: update if existed, create if not existed
@@ -306,7 +317,8 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
         this._createOrUpdateLine(positions, this.dragHandler.draggedObjectInfo.lines, {
             status: "completed",
             color: this.stateManager.getColorState("line"),
-            interactive: true
+            interactive: true,
+            listeners: this.#polylineListeners
         });
 
         // -- Finalize Label Graphics --
@@ -331,97 +343,7 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
     /**********
      * HELPER *
      **********/
-    /**
-     * Creates a new polyline or updates an existing one based on positions.
-     * Manages the reference within the provided polylinesArray.
-     * @param {{lat: number, lng: number}[]} positions - Array of positions to create or update the line.
-     * @param {L.polyline[]} polylinesArray - The array (passed by reference) that holds the polyline instance. This array will be modified. Caution: this is not the polylineCollection.
-     * @param {Object} [options={}] - Options for the line.
-     * @returns {L.polyline | null} The created or updated polyline instance, or null if failed.
-     */
-    _createOrUpdateLine(positions, polylinesArray, options = {}) {
-        // Validate positions input
-        if (!Array.isArray(positions) || positions.length < 2 || !positions[0] || !positions[1]) {
-            console.warn("_createOrUpdateLine: Requires an array with at least two valid positions.");
-            return null;
-        }
-        // Validate polylinesArray input
-        if (!Array.isArray(polylinesArray)) {
-            console.warn("_createOrUpdateLine: polylinesArray argument must be an array.");
-            return null;
-        }
 
-        // Default options
-        const {
-            status = "pending", // Default pending status
-            color = this.stateManager.getColorState("move"),
-            id = `annotate_${this.mode}_line_${this.measure.id}`,
-            interactive = false,
-            ...rest
-        } = options;
-
-        let lineInstance = null;
-
-        // -- Update existing polyline --
-        if (polylinesArray.length > 0) {
-            lineInstance = polylinesArray[0]; // Get the reference from the array
-
-            // Simplified Check: Assumes if exists, it's valid.
-            if (!lineInstance) { // Check if the retrieved reference is truthy
-                console.warn("_createOrUpdateLine: Invalid (null/undefined) object found in polylinesArray. Attempting to remove and recreate.");
-                polylinesArray.length = 0; // Clear the array to trigger creation below
-                // Fall through to the creation block (lineInstance is null)
-            } else {
-                // -- Handle Polyline Visual Update --
-                // Assumes lineInstance is a valid Polyline if it exists
-                lineInstance.setLatLngs(positions); // Update path of the line, requires L.latLng[]
-                lineInstance.setStyle({ color: color }); // Update color
-
-                // Update lineInstance interactive attribute
-                const oldInteractiveState = lineInstance.options.interactive;
-                // Compare the old with current interactive state, only update interactive if different
-                if (oldInteractiveState !== interactive) {
-                    // Update the interactive
-                    lineInstance.options.interactive = interactive;
-                    // Refresh the layer to apply the new interactive state. 
-                    if (this.drawingHelper && typeof this.drawingHelper._refreshLayerInteractivity === 'function') {
-                        this.drawingHelper._refreshLayerInteractivity(lineInstance);
-                    }
-                }
-
-                // -- Handle Metadata Update for new and existed line --
-                Object.assign(lineInstance.feature.properties, {
-                    status,
-                    positions: positions.map(pos => ({ ...pos })), // Store positions copy
-                    ...(id && deconstructIdForMetadata(id)), // Deconstruct id for metadata
-                })
-                lineInstance.feature.id = id; // Update the id on the feature
-                lineInstance.id = id; // Update the id on the line instance
-            }
-        }
-
-        // --- Creation new polyline ---
-        // This block runs if polylinesArray was empty OR if the existing entry was invalid (!lineInstance was true above)
-        if (!lineInstance) { // Check if we need to create (either initially empty or cleared due to invalid entry)
-            lineInstance = this.drawingHelper._addPolyline(positions, {
-                color,
-                id,
-                status,
-                interactive,
-                ...rest
-            });
-
-            if (!lineInstance) {
-                console.error("_createOrUpdateLine: Failed to create new polyline instance.");
-                return null; // Return null if creation failed
-            }
-
-            // -- Handle References Update --
-            polylinesArray.push(lineInstance); // Push the new instance into the referenced array
-        }
-
-        return lineInstance; // Return the instance
-    }
 
     /**
       * Create or update the label.
@@ -451,13 +373,6 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
 
         const distance = calculateDistance(positions[0], positions[1]); // calculate distance
         const formattedText = formatMeasurementValue(distance, "meter"); // Format the distance value
-        const middlePos = calculateMiddlePos(positions); // calculate label position
-
-        if (!middlePos) {
-            console.warn("_createOrUpdateLabel: Failed to calculate middle position.");
-            // Return distance but null instance if middle position calculation fails
-            return { distance, labelInstance: null };
-        }
 
         let labelInstance = null;
 
@@ -470,37 +385,14 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
                 console.warn("_createOrUpdateLabel: Invalid object found in labelsArray. Attempting to remove and recreate.");
                 labelsArray.length = 0; // Clear the array to trigger creation below
             } else {
-                // -- Handle Label Visual Update --
-                labelInstance.setLatLng(middlePos); // update position
-
-                // Create HTML element for label content
-                const contentElement = document.createElement('span');
-                contentElement.style.color = color;
-                contentElement.textContent = formattedText;
-
-                // Set the content of the label
-                labelInstance.setContent(contentElement); // update content
-
-                // Update interactive state
-                const oldInteractiveState = labelInstance.options.interactive;
-                // Compare the old with current interactive state, only update interactive if different
-                if (oldInteractiveState !== interactive) {
-                    // Update the interactive
-                    labelInstance.options.interactive = interactive;
-                    // Refresh the layer to apply the new interactive state. 
-                    if (this.drawingHelper && typeof this.drawingHelper._refreshLayerInteractivity === 'function') {
-                        this.drawingHelper._refreshLayerInteractivity(labelInstance);
-                    }
-                }
-
-                // -- Handle Metadata Update for new and existed label --
-                Object.assign(labelInstance.feature.properties, {
+                // Update label visuals and metadata
+                labelInstance = this._updateLabel(labelInstance, positions, formattedText, {
+                    id,
                     status,
-                    positions: positions.map(pos => ({ ...pos })), // Store positions copy
-                    ...(id && deconstructIdForMetadata(id)), // Deconstruct id for metadata
-                });
-                labelInstance.feature.id = id; // Update the id on the feature
-                labelInstance.id = id; // Update the id on the label instance
+                    color,
+                    interactive,
+                    ...rest
+                })
             }
         }
 
@@ -508,6 +400,7 @@ class TwoPointsDistanceLeaflet extends MeasureModeLeaflet {
         if (!labelInstance) {
             labelInstance = this.drawingHelper._addLabel(positions, distance, "meter", {
                 id,
+                color,
                 status,
                 interactive,
                 ...rest
